requirements.txt
asn1crypto==0.23.0
boto3==1.4.4
botocore==1.5.62
certifi==2017.7.27.1
cffi==1.11.0
chardet==3.0.4
click==6.6
cryptography==2.0.3
docutils==0.12
enum34==1.1.6
futures==3.0.5
idna==2.6
ipaddress==1.0.18
jmespath==0.9.0
lxml==4.0.0
pyaml==15.8.2
pycparser==2.18
pyOpenSSL==17.3.0
python-dateutil==2.5.3
python-lambda==2.1.0
PyYAML==3.11
requests==2.18.4
s3transfer==0.1.11
six==1.10.0
urllib3==1.22
